restructure (model -> controller); bug fix (song_list update)
